Add routing primitives (#1)
This change adds routing primitives by means of a Router module which exposes
the necessary building blocks to translate from urls to endpoint and build links
to endpoints. At the core is the page variant which assumed along screens as
coarse unit to route to. These pages can then be used to vary on them in the
display logic of the main app. In turn the pages are used to construct working
links to build up navigations and references throughout the app.

* implement Router module
* add Router type declarations
* use stop-gap Generic page
* build minimal navigation component