Implement project sub navigation (#2)
As an exercise, and we have to support multiple functional contexts
within a project, this change explores how a sub navigation with
associated routes can look like. It leaves the main composition to the
App component and enables different sub pages with a prop passed to
a page. Along the way it entirely hides the internal use of
ReasonReactRouter in the Router module by exposing `currentPage` and
instead of returning a raw string, that can be used as a link, it
returns a function that can be invoked when the new pushState should
take effect.

* implement Project component
* use props to enable Project sub page routing
* show simple project navigation which shows active "tab"
* use `currentPage` as the main routing interface
* move from raw strings to `navigate` functions to be used by components
which want to build links