Introduce app store (#9)
To manage data, side-effects and how their result flow through the app,
this change-set introduces a store which is widely available to pages
and components. Roughly following The Elm Architecture there is a
central reducer which dispatches to sub-reducers based on actions
expressed as variants. To show how to apply these we introduce a simple
project registration.

* implement app level reducer
* build store and provider abstraction based on React context api
* implement sub reducer for projects
* show simple project registration

*****************************************************************************
* Add project registration page
Introduce project registration for the user to add their own project to
the network. In the simplest form we just ask a bare minimum of
information and have the library interact with whatever backing store is
responsible for project handling.

* introduce new RegisterProject page
* use context to keep track of state and re-render dependent components
* Meditate on redux like store
* Expand store infrastructure
* Enable redux debugging enhancing
* Handle data fetching in thunk
* Consolidate data provision in Source
* Seperate project store parts into own module
* Improve code split to avoid import hell
* Add register thunk
* Fix project list items (#15)
List items in React are required to have a unique key assigned for the
underlying diffing algorithm to work properly and compute which elements
need redrawing. We also missed the `li` for the enclosing `ul`, just
a small semantic fix on top.

* add key prop to project list items
* wrap list item in `li`
* Store and return updated projects
* Annotate source
* Update app/src/Source.rei
Co-Authored-By: Rūdolfs Ošiņš <rudolfs@osins.org>
* Update app/src/Source.rei
Co-Authored-By: Rūdolfs Ošiņš <rudolfs@osins.org>
* Update app/src/Source.rei
Co-Authored-By: Rūdolfs Ošiņš <rudolfs@osins.org>
* Update app/src/pages/PageProjects.re
Co-Authored-By: Rūdolfs Ošiņš <rudolfs@osins.org>
* Update app/src/Source.rei
Co-Authored-By: Rūdolfs Ošiņš <rudolfs@osins.org>
* Change reductive dependency to use owned fork