Enable rollup pipeline for electron scripts (#43)
To ensure a consistent language variant and style we use the same build
pipeline for our electron scripts. We keep the main rollup.config.js and
import an rollup.app.js and rollup.electron.js one. This enables us to
build configs exlcusively for either and combine them when necessary.
The plugins used for the electron one assume a node environment (where
electron scripts are executed).