Code browsing (#55)
* Draft GraphQL queries
* Update schema with more types for Git related info
* Fix some syntax in the schema
* Add fake radicle backend
* Add source tree to sidebar
This gets data from the fake radicle GraphQL backend and shows a dummy
source tree in the sidebar under the "Source" item.

* Fix GraphQL types
Types are capitalized in GraphQL.

* Add branch/tag/commit dropdown
* Revert extract routes
After thinking about it, it doesn't make sense to group navigation item
data into routes. There has to be a better way.

* Centralize navigation logic
This is the second attempt at getting the navigation logic right. Now
all the logic around parsing and matching paths lives in one place:
`path.js`. It had to be separate from `routes.js` to avoid circular
dependencies.

* Add profile avatar active state
* Make the fake backend work with absolute paths
* Disable fake backend debug messages by default
* Split routes from paths
Wasn't a good idea after all.

* Re-structure component nesting
* Make folder and file links work
* Remove unneeded code
* Fetch available heads from backend
* Fix wording
* Style source code component
* Fix line-numbering
Off by one, lol!

* Prevent user from selecting line numbers
* Don't try to render binary files
* Show file path
* Remove debugging statement
* Simplify project layout
Do layout styling in the Project page, as it is the top-level component
for all the project pages.

* Fix small button radius
* Add naive dropdown styling
* Use dropdown component in Source page
* Add CommitTeaser component
* Add back error logging
For some reason prettier strips it out.

* Start with main window maximized
All our design prototypes assume that the app takes full-screen
dimensions.

* Add style overrides to all icons
* Add directory listing markup and style
* Tweak folder caret hover state
* Add project breadcrumbs
* Improve schema to incorporate revisions
Also added an example of how to get the latest commit for a path.

* Incorporate feedback
* Add sha attribute to Commit type
* Rename head -> revision
* Rename cat -> blob
* Rename ls -> tree
* Clean up logging
* Simplify schema to accomadate needs for git interactions
* Pass projectId via context
* Split out revision selector
* Fix spelling of schema field
* Show file list or file content depending on object type
* Display files up to 1MB
* Visually mark active file/folder
* Fix project repository paths
This allows us to point our git qgl wrapper at any repository.

* Show commit data in source browser
* Show last commit data for trees
* Update schema
* Rename leaves -> entries
Because we're not dealing with a real git tree object.

* Add more documentation for GraphQL api
* Fix redundancies in paths
* Make default project revision a constant
* Use GraphQL enum to distinguish tree/blob types