Update to Registry memory client (#61)
This is the groundhogday update we have to perform to keep up with the
client changes in the registry. It comes with the added bonus that we
adjust to the new understanding of IDs in the Registry from a hash to
`(ProjectName, ProjectDomain)` tuple. Wich in turn led to quite some
changes in the app to consume and provide these properly.

* Adapt major parts to memory client
* Finalise integration
* Please clippy
* Bump juniper dependencies
* Use correct flag for source
* Make juniper_codegen an explicit dependency
* Start fixture implementation
* Handle IDs in graphql API correctly
* Pass registry memory flag consistently
* Disable debug flag in fake git backend
* Remove trailing spaces
* Remove old ProjectId implementation
* Fix breadcrumb and sidebar