Integrate radicle-surf to satisfy CoCo API endpoints (#64)
This change introduces an implementation of the API drafted for #55
powered by radicle-dev/radicle-surf. A couple of adjustments had to be
made to surf along the way. Most notably we needed to reexport git2 to
have access to some lower level types. We track types and their info we
depend on in radicle-dev/radicle-surf#41 to remedy that. We also settled
on git-platinum as the body of fixtures to be used in our tests. The
infrastructure to test GraphQL queries has been drastically improved in
this patch and is expected to be more exhaustive with additions, as well
as the discovery of edge cases.

At the time of this writing there is a incosnistency in how
`last_commit` behaves, it is tracked in radicle-dev/radicle-surf#40.
This means the API will always return the latest commit for the current
history, indiscriminate if it touched the path in question - Resulting
in errors in the ui integration tests.

* implement branches query
* implement tags query
* implement commit query
* implement blob query
* implement tree query
* integrate API in the UI
* use radicle-dev/git-platinum for test fixtures
* use latest compatible toolchain (dictated by radicle-dev/radicle-registry)
* pull in radicle-dev/radicle-surf as dependency

Resolves MON-10