Releases (#49)
Initially we wanted to fully automate releases and build artifact
uploads, but as @kim pointed out it might be a security hazard and also
might mislead people into thinking this is an official production
release. So until we have the proper infrastructure in place to securely
automate releases, we'll do them manually.

It's worth mentioning that the releases are meant for our own internal
evaluation only as the app is still in prototype stage. For this I added
a warning in the readme.

* Configure electron-builder release
* Build both platforms in parallel
* Add author and description
* Explicitly spell out Linux package types
* Build macOS only for master
* Update README
* chore(release): 0.0.2
* Remove publish steps
* Extract development related documentation
* Fix link
* Fixup
* Fixup
* Improve wording
* Update contact info
* Automate releases
* Fixup
* Fixup
Co-authored-by: Rūdolfs Ošiņš <rudolfs@osins.org>