refactor(ui): readme fetch and rendering (#335)
This improves the way the Readme is fetched and rendered and gets rid
of a lot of nested conditionals. Achieved through the extension of
Remote with an optional context it can create, which is sub-optimal but
the best approach at the moment. We also introduce store chaining,
where one store is dependent on the successful remote fetch of another.

Follow-up to #311


* Clean up component props
* Implement readme as chained store
* Rework chaining into Promise
* Fix inconsistencies
* Extend Remote with optional context
* Add missing component file
* Fix param and revision reactivity