refactor(ui): consolidate to ts (#357)
A set of consolidations to make the codebase more uniform and have
almost all non Svelte ui code in typescript. Will leave hotkeys as a
follow-up as it currently relies on the router which can't be imported
as a normal module as it is implemented as a Svelte component for
whatever reason. I do have an idea of how to address that, but it would
bloat the scope here.

* Remove unnecessary compile outputs
* Migrate path to ts
* Migrate commit mocks to src
* Remove unused hash module
* Migrate notification to ts
* Migrate config to ts
* Remove the need for native types module
* Fix store access
* Remove unused org mocks
* Name project creation screen consistently
* Correct compiled js glob
* Name identity creation fragments consistently
* Fix notifications emission and visuals
* Render NotificationFaucet app globally