feat(proxy): improve session
Addresses the sub-optimal session handling, by actually serialising the
object itself into the store. This allows us to also keep the user
handle that was registered around and feed it to the org listing. That
means that all workflows that require local understanding of what
Registry entities are associated with the current session should work.

* serialise session object into store
* store registered user handle in session
* use user handle to query session org list
* improve internal registry API type coherence
* put a bunch of serialisations in the right place

Fixes #378