fix(proxy): only fetch commit for root tree (#405)
This fixes a long standing performance issue with the file browser. There's still some room for improvement (the file list still doesn't load *instantly*), but I think that will require more involved work, such as a custom endpoint. For now, this makes it useable at least.

Fixes #350 

The issue was two-fold: the `tree::last_commit` function from `surf` currently has to go through the entire file (or tree) history to return the last commit. And we were calling this function for every folder in the tree.

The simple fix was to only call that function for the root folder, since we don't actually use the information for the sub-folders currently. We should also eventually make the `last_commit` function faster.