refactor(ui): move all repository browsing state into store (#564)
So far part of the source browsing state was in the router and another
part in the global store. With this PR all the state will live in the
Typescript/Svelte store.

We don't gain anything by having this state in the browser's
window.location as it isn't presented to the user. And if we will want
to present it to the user it's going to be easier to just grab it from
the store.

This is in preparation of the UI changes for #443.