Complete Refactor of Peer
The main goal here is to move from using Peer, which wrapped the PeerApi
in an Arc + Mutex, to just having functions that take &PeerApi.
I created a new module to house these PeerApi operations -- similar in
vein to source housing Browser operations.
This means that the http endpoints take PeerApi wrapped in Arc + Mutex
and avoids the indirection.
This also meant that we need to avoid using PeerApi in functions that
involve a async function being called. If that's necessary we need to
pass an Arc + Mutex PeerApi.

The result of this is that we avoid this locking inside our peer
operations and can freely use it.