Direct radicle-surf dependency
We have begun to publish radicle-surf to crates.io again. In this change
we use this method for retrieving radicle-surf rather than through
librad.

The major difference here is that we can't simply create a Browser from
the Storage/Repo anymore. Instead, we must use the Paths to get the
git_dir and setup the Browser that way -- also calculating the Namespace
for the Project's RadUrn.