chore(proxy): update librad (#616)
* Updating to latest changes in librad.

The secret key is not available to us anymore through the PeerApi. So
instead, we have to pass in the SecretKey where needed. This means that
we should use the Keystorage for getting the key from file and passing
it into the functions that require it -- not holding onto the key
indefinitely.

When creating a project, we need to ensure there's a certifier, which is
initially the owner of the project. This means we avoid the use of
fake_owner and create the initial User in the monorepo.

* Need to nuke the keystore now as well to ensure we get the right keys

* Remove fake_owner

* Add test case for browsing a commit header