feat: implement nuke command
Developers often are unable to start up the app because the state on
disk is not compatible with code versions or it has been corrupted in
other ways. To remedy that and provide a drastic tool to get uninitated
unstuck without the knowledge of the on disk fs layout, this change-set
introduces `yarn nuke`.

Closes #394