Update surf to 0.4.0
This means that we have structured revisions now. For some of the
endpoints, this is a minor change. What's passed in can be converted to
a Rev. For blob and tree, we need a way to get the structured input for
the revision -- instead of guessing from a string. The structure can be
a Tag, Sha, or a Branch (with an optional remote PeerId).