Have tree and blob take JSON body
We were running into an issue that the warp url decoding was munging our
json strings for Revision. We would get '\\\' which would then fail to
deserialise to Revision (can be found by logging the output of warp
using RUST_LOG=warp=debug).

Instead we take the Revision as part of the body of the request where
we're safe to (de)serialise.

To confirm this works we added two new tests that look at the 'dev'
branch, to ensure we're not falling back to the 'master' branch.