chore: update surf to 0.4.0 (#654)
* Update surf to 0.4.0

This means that we have structured revisions now. For some of the
endpoints, this is a minor change. What's passed in can be converted to
a Rev. For blob and tree, we need a way to get the structured input for
the revision -- instead of guessing from a string. The structure can be
a Tag, Sha, or a Branch (with an optional remote PeerId).

* Have tree and blob take JSON body

We were running into an issue that the warp url decoding was munging our
json strings for Revision. We would get '\\\' which would then fail to
deserialise to Revision (can be found by logging the output of warp
using RUST_LOG=warp=debug).

Instead we take the Revision as part of the body of the request where
we're safe to (de)serialise.

To confirm this works we added two new tests that look at the 'dev'
branch, to ensure we're not falling back to the 'master' branch.

* Use serde_qs

We combine the 'raw' query method in warp with the serde_qs library so
that we can (de)serialise more complicated query structures for
endpoints.

* Add RevisionQuery type for source browsing
* Add logging to with_qs
* Use typed revision in UI
* Fix tests

Fixing some of the tests that weren't fixed in a previous WIP commit.
Also fixing the URL de/encoding. The serde_qs library expects
non-encoded '[' and ']' characters. So we use the percent-encoding
library to decode the raw string first and then deserialise using the
serde_qs library.

* Pass only revision to commits endpoint
* Pass in Branch to commits query
* Shareable Identity should be <handle>@<peer_id>

The RadUrn identifies a user's User profile in a monorepo sense. The
shareable identifier, however, should be the PeerId. This is highlighted
when we were looking at revisions and we need the PeerId to inspect the
remote.

* Better check for peer is not self
* Manually compare objects for better readability

Co-authored-by: Rūdolfs Ošiņš <rudolfs@osins.org>