Localise the revisions function in coco/source.rs
Moving the revisions function into source so that it only operates over
a Browser. The peer logic gets moved to handler and the information gets
passed down.

As well as this, we extend the branches endpoint so that it can take a
PeerId if we want to look at a peer's list of branches. We test this in the revisions test.