feat(proxy): peer branches (#678)
* Use typed RadUrn params

* Localise the revisions function in coco/source.rs

Moving the revisions function into source so that it only operates over
a Browser. The peer logic gets moved to handler and the information gets
passed down.

As well as this, we extend the branches endpoint so that it can take a
PeerId if we want to look at a peer's list of branches. We test this in the revisions test.

* Fmt all the code

* Make McClippy happy

* Fix docs

* Remove info log

Co-authored-by: Rūdolfs Ošiņš <rudolfs@osins.org>

* Rename PeerQuery to BranchQuery

* Refactor UserRevisions

* Renamed UserRevisions to Revisions.
* Removed Identity and replaced with peer_id, user fields.
* Parameterise over P and U so that we don't import librad modules.

* Parameterise over Revision too

* Have a Reply version of Revisions that holds the Identity

* Update proxy/src/http/source.rs

Co-authored-by: Alexander Simmerl <a.simmerl@gmail.com>

* Document Revisions output

* Fix docs

Co-authored-by: Rūdolfs Ošiņš <rudolfs@osins.org>
Co-authored-by: Alexander Simmerl <a.simmerl@gmail.com>