Build project path
Before this change we would accept the path provided as the destination
for the project to be set up. But since the user inputs a name it would
be a better UX if they picked the base directory, i.e. some workspace,
and the project's name serves as the directory.

However, if the user is selecting a directory that already ends in the
name, then we should just take that as the path.