Cleanup and test checkout
We move the checkout functionality into project.rs since it doesn't need
to know about the PeerApi.
We create a data type for the checkout to indicate what information is
needed to execute the checkout.
Finally, we add an integration test so that we can confirm the checkout
works with git-rad-remote. Note that this needs to be an integration
test in order to have this executable available to us.