Project Creation
This commit splits project creation off from the peer logic. The change
captures that there are two ways to create a project in Upstream.
There's a completely new repository, or an existing repository.
In the former we expect slightly different inputs, where we build the final
directory from the supplied path and name.
In the later, we expect the existing directory to already be a git
repostiory and the last portion of the path is the name of the project.
The intention now needs to be set when creating a project via the HTTP
endpoint, instead of making guess work.

Included in this is some general refactoring of the control code which
seemed useful to have for testing.