Update frontend
We match the API call on the front end to specify whether we want to
create a new or existing repository.
If existing is selected then we take the name of the folder as the name
of the project and disable the input box.
There's no behavioural change for creating a brand new project.