Clean up checkout
Turns out we didn't need to call the git commands from the command line.
All we needed was to register the git transport and this is the same
thing. This means we can use git2 to execute actions. So instead of
implementing the checkout via git clone as a Command, we use
git2::RepoBuilder.