Delete an existing remote
We added more test cases to ensure that when creating from an existing
repository we don't get any errors. We have to delete and re-create the
remote if there is an existing rad remote and the URLs don't match.
This mismatch in URLs causes issues with finding the correct branch to
browse.