fix(proxy): rad remotes (#781)
* Allow second creation

We don't want to fail on an existing remote. Instead, we'd like to still
create the repo since it's valid if it is existing. However, a change
needs to be made to check that fetchspecs are still valid.

* Attempting to get Remote working

Trying to get Remote working where it points to a librad::LocalUrl.
Unfortunately, it's not. I think this is due to some errors with the
credential helper.

* Formatter, Clippy, Fix tests

* Rework coco project module

The checkout logic wasn't necessarily part of the proxy project, so I
decided to move it to the recently added coco::project module. To
further separate concerns, I split this module into submodules of
checkout and create. This required some refactoring of imports and at
the same time I refactored the errors into separate enums.

* WIP: Use "git push" via a command process

To get access to the rad remote helper, we can call git push via a
process and pass in the password credential helper (a stop-gap
solution).

* WIP: Trying to get remote working without rad remote helper

* Update librad SHA

Updating librad allows us to have concurrent tests working. With some
tweaks to doing the pushes without the refspecs we get all the tests
passing and can clean up some of the previous efforts.

* Clean up params and errors

* Clean up

* Formatter

* Clean up checkout

Turns out we didn't need to call the git commands from the command line.
All we needed was to register the git transport and this is the same
thing. This means we can use git2 to execute actions. So instead of
implementing the checkout via git clone as a Command, we use
git2::RepoBuilder.

* Just check that we have a rad remote

* Update dependencies

* Formatter

* Fix typo temdir -> tempdir

Co-authored-by: Rūdolfs Ošiņš <rudolfs@osins.org>

* Fix: missing matching ' in debug message

Co-authored-by: Rūdolfs Ošiņš <rudolfs@osins.org>

* Replace docs that were deleted by accident

* Move RAD_REMOTE to config

* Update librad dependency and remote creation

It turned out that we didn't need to setup any fetch or push refspecs.
Instead we just need to create the remote with the URL and then
everything works from there on.

* Remove the check for managed

When creating an existing project based off of one that is already
managed by us, we don't want to fail. We should allow this.

* fixup! Remove managed field

* Set the upstream branch to use

We set the rad remote as the upstream for the default branch.
Changing the existing repo test to make sure it works correctly.

* Update Cargo.lock

* Fix test

* Remove test guarding against existing remote

* Delete an existing remote

We added more test cases to ensure that when creating from an existing
repository we don't get any errors. We have to delete and re-create the
remote if there is an existing rad remote and the URLs don't match.
This mismatch in URLs causes issues with finding the correct branch to
browse.

* Ensure the default branch is pushed

We want to make sure that default branch is always pushed in any of the
cases of setting up the remote. If it isn't a downstream Browser will
have a bad time at trying to get at references.

* Formattings

* Revert change on setting rad/self

* Update warning message typo: missing '

Co-authored-by: Alexis Sellier <alexis@monadic.xyz>

* Remove uneeded RAD_HOME setting

* Appease clippy

Co-authored-by: Rūdolfs Ošiņš <rudolfs@osins.org>
Co-authored-by: Alexander Simmerl <a.simmerl@gmail.com>
Co-authored-by: Alexis Sellier <alexis@monadic.xyz>