Rework listing projects
The idea here is to differentiate between the different types of
projects. We have Tracked and we have Mine. When we list them, we get
back all of them but throw them into their own buckets, and tag tracked
ones with a new type.
Then in the endpoint call we can take a query parameter designates which
type of projects the caller is looking for.

WIP: Move call to list_projects_for_user into it's own endpoint call