Initial Draft of Request & Waiting Room
A request is modeled as the data it needs to fulfill a request and the
metadata that it encounters along its journey. We use a PhantomData
marker to ensure that transitions between our states are valid.

The waiting room is then modeled as a collection of these requests keyed
by their RadUrn. The error handling is done within the waiting room
since it needs to hide away what kind of request we had.

There's a single happy path test to get us going.