Represent intermediate transitions at type level
We introduce the intermediate states, Found, Cloning, and Cloned at the
type level so that we have more fine grained transitions. This means we
can't have just PhantomData as the state.

This introduces those changes and transitions while fixing the
end-to-end test.

However, there is still work to be done on the business logic that
happens at each state and when we need to transition to a failure state
such as canceled or timed out.