feat: introduce RADICLE_UPSTREAM_EXPERIMENTAL flag
This is a feature flag that is consumed by the app as `isExperimental`,
signalling whether we are running the app with the experimental mode on,
which toggles experimental/under construction features. This flag is
only allowed in development mode.
A new yarn command is added, `yarn start:public`, which starts the app
in public mode, i.e, hiding all the experimental features. Such run
simulates a production preview while running in development mode.