Organising shared functionality through traits
We capture shared functionality: Having peers and be able to cancel a
request, and we organise them in traits. This allows us to give
instances of these traits to the states that they are relevant for. This
then means we can write generic transitions for Request.

Since these traits are very particular they are sealed by this module.