feat(proxy): differentiate tracked projects from my projects (#866)
* newtype for Tracked projects, query param

* Move my_projects into api/project

To ensure that we don't bloat the Api struct, we move my_projects into
the api workspace. This is because it can already be defined in terms of
the components of Api, and it's a domain specific interpretation of what
"my project" is.

* Rework listing projects

The idea here is to differentiate between the different types of
projects. We have Tracked and we have Mine. When we list them, we get
back all of them but throw them into their own buckets, and tag tracked
ones with a new type.
Then in the endpoint call we can take a query parameter designates which
type of projects the caller is looking for.

WIP: Move call to list_projects_for_user into it's own endpoint call

* Add camelCase pragma and clean up params

* trying to impl IntoIterator for Projects

* compiles but tests fail

* separate endpoints for tracked/untracked

* get all projects for given identity in identity module

* add user query param to tracked/ endpoint

* update ui endpoints

* separate endpoint for peer projects

* small fixes

* update test name

* integration tests failing

* Surface error lines from proxy

* skip tests that depend on peers

* remove tests related to tracked projects bc we cant actually test that

* list_for_user test update

* Idiomatic Iterators

Introduce and Iter and IntoIter struct for Projects. This allows us to
iterate over the projects without any clones.

* FETCH_User -> FETCH_USER

* Fix sentence to say "current user"

* Apply suggestions from code review

Co-authored-by: Nuno Alexandre <hi@nunoalexandre.com>

* Update documentation and naming

We want to be very explicit about what kind of user we're talking about,
i.e. a user that is not us, and what kind of projects we're returning
(in this case we're not differentiating).

* Fix naming of function calls

* Fix documentation links

Co-authored-by: FintanH <fintan.halpenny@gmail.com>
Co-authored-by: Alexander Simmerl <a.simmerl@gmail.com>
Co-authored-by: Nuno Alexandre <hi@nunoalexandre.com>