Correct the all_failed method
We first defined all_failed with &mut but this is too much power for
checking the HashMap.
Instead, we give Found and Cloning Deref impls where the targets are
their respective HashMaps, and then we can define all_failed in terms of
Deref.
This means that the API should be accessible via the Deref of Request
followed by a call to all_failed.