feat(proxy): waiting room requests (#903)
* Initial Draft of Request & Waiting Room

A request is modeled as the data it needs to fulfill a request and the
metadata that it encounters along its journey. We use a PhantomData
marker to ensure that transitions between our states are valid.

The waiting room is then modeled as a collection of these requests keyed
by their RadUrn. The error handling is done within the waiting room
since it needs to hide away what kind of request we had.

There's a single happy path test to get us going.

* Implement exponential back off helper

* Implement From impl for Request -> Gossip

* Add Error type and organise modules

Moving the waiting room into its own module to separate the boundary
that it's a separate entity to request.

This makes it easier to define the Error types locally and replace the
() errors with these domain specific errors.

* Add Strategy for selecting next

We add a Strategy enum so that the caller can pick which kind of
strategy they want when selecting the next Request in a Created state.

* Represent intermediate transitions at type level

We introduce the intermediate states, Found, Cloning, and Cloned at the
type level so that we have more fine grained transitions. This means we
can't have just PhantomData as the state.

This introduces those changes and transitions while fixing the
end-to-end test.

However, there is still work to be done on the business logic that
happens at each state and when we need to transition to a failure state
such as canceled or timed out.

* exponential_backoff over just one attempt number

* Update semantics

We make more transitions explicit between request states and allow to go
back to previous states. We also keep track of how a clone went for a
given PeerId.

Left some inline comments as reminder for what's next.

* Organising shared functionality through traits

We capture shared functionality: Having peers and be able to cancel a
request, and we organise them in traits. This allows us to give
instances of these traits to the states that they are relevant for. This
then means we can write generic transitions for Request.

Since these traits are very particular they are sealed by this module.

* Add sealed module

* Move timed_out to generic Request

* Clean up transitioning

We have some cases where it's easy to define the functions over a list
of states.

We also clean up some of the functionality and code.

* Module split

The code is easier to browse if we split the state and the existential
code into sub-modules.

* Bump either version

* Add a filter function to define next and ready

* Add De/Serialize impls

Using deriving for Serialize and Deserialize for all the types so that
we can ultimately serialize and deserialize a WaitingRoom.

* Type Safe Queries + Clones

Since I clearly made a mistake that was caught in code review, I won't
let it happen again. It shall be caught at compile time! usize no more,
only Queries and Clones!

* Move Attempts, Queries, and Clones into states module

* Capture time out logic

We should time out when we are in one of the three states IsRequested,
Found, or Cloning. This is when we check if we've reached the max amount
of attempts and either return the timed out request or the successful
transition.

* Introduce a Config for WaitingRoom

Instead of individual fields in the WaitingRoom we combine them into a
single Config type. Its Default impl uses the const values.

* Generic next on Strategy

We can define Strategy's next function over a generic iterator where we
can extract keys. This will make it easier for testing the behaviour
creating an iterator of easy to create items, rather than requests.

* Add more testing of expected behaviour

* Remove exponential backoff

Since this function depends on nothing in our domain here, we believe it
can and should be defined by an upstream component that uses request.

* Document and clean up

Documenting the Request API.
Conistently naming the `PeerId` variable to `peer`.
Noticed that we don't move a `peer` into `InProgress` when cloning so we
fixed that in this commit.

* Unify waiting room functions

To hide the difference between state transitions and looping states, we
unify the waiting room functions as a black box.
We also convert time outs into a waiting room error.

* Document state types and traits

* Document SomeRequest

* Document and clean up

* Fix tests & Run Clippy

* Have a way to transition back from Found to IsRequested

* Fix typos in documentation and error message

Co-authored-by: sarah olimpia <sarahscott@users.noreply.github.com>
Co-authored-by: Rūdolfs Ošiņš <rudolfs@osins.org>

* cargo fmt

* Move proptest macro to test suite

* Add tracking issue of clippy bug

* Move module documentation into modules

* Move the MAX_* constants to waiting room, and give them a more fair
value.

* Rename failed to cloning_failed and fix its documentation

* Add Deref impl for S on Request.

This allows users of the API to access a reference to the underlying
state.

* Correct the all_failed method

We first defined all_failed with &mut but this is too much power for
checking the HashMap.
Instead, we give Found and Cloning Deref impls where the targets are
their respective HashMaps, and then we can define all_failed in terms of
Deref.
This means that the API should be accessible via the Deref of Request
followed by a call to all_failed.

* Add found_failed method to waiting room

We added failed for the Found state in the Request API, but missed
adding it to the WaitingRoom API.

* Rename list to iter and return the pairs

* Use self directly rather than deref

* Extend cannot_create_twice test

To show that the test is working as expected we advance the request
state, attempt to create a request under the same URN and see that we
get back the advanced request.

* Requested can know about peers.

We change the Requested state to hold a HashMap of peers. This is so
that when we fail back to this state then we don't forget the history of
clone requests for peers.

* Add join to Status and fix semantics

To ensure that we are not changing the Status of a peer clone by
accident we introduce join on Status which defines Failed as an upper
bound.

* Fail path from Cloning to Requested

The Request, if failing in the Cloning state, should fail back to the
Requested state if all the peers have been exhausted. If they haven't
then it should only fail back to Found.

* Rename create to request

We rename create to request since it does not always create, and
conditionally returns a request. We also update the docs to better
reflect this.

* Ok, Clippy

* Rename Is* states

* Remove TODO

While it is weird, the caller of the waiting room/request API has to
hold themselves accountable in this case. If they say a peer failed then
we mark it as failed, whether it existed in the map or not.

* Introduce elapsed

Introducing a method that checks the elapsed time between a request and
provided timestamp.

* Use general filter function instead

We introduce a filter function that bases the filtering criteria off of
the RequestKind and the delta of elapsed time based on a provided
timestamp. This deprecates the next and ready functions.

* Bye, bye Strategy and bye, bye proptest

* Ok, FMT

* Fix typo

Co-authored-by: Nuno Alexandre <hi@nunoalexandre.com>

* Remove extraneous docs

Co-authored-by: Nuno Alexandre <hi@nunoalexandre.com>

* Rename RequestKind to RequestState

* Rework docs on Status

Co-authored-by: sarah olimpia <sarahscott@users.noreply.github.com>
Co-authored-by: Rūdolfs Ošiņš <rudolfs@osins.org>
Co-authored-by: Nuno Alexandre <hi@nunoalexandre.com>